folder structure changes, namespace files instead of folders
complaince articles have nothing to do with mines, so are pulled. tests moved to match app folders.
