+++ python-backend/app/__init__.py
@@ -6,15 +6,16 @@ from sqlalchemy.exc import SQLAlchemyError
 
 from flask_jwt_oidc.exceptions import AuthError
 
-from app.api.parties.namespace.parties import api as parties_api
-from app.api.mines.namespace.mines import api as mines_api
-from app.api.download_token.namespace.download_token import api as download_token_api
-from app.api.users.namespace.users import api as users_api
-from app.api.search.namespace.search import api as search_api
-from app.api.reporting.namespace.reporting import api as reporting_api
-from app.api.variances.namespace.variances import api as variances_api
-from app.api.incidents.namespace.incidents import api as incidents_api
-from app.api.now_submissions.namespace.now_submissions import api as now_api
+from app.api.compliance.namespace import api as compliance_api
+from app.api.parties.namespace import api as parties_api
+from app.api.mines.namespace import api as mines_api
+from app.api.download_token.namespace import api as download_token_api
+from app.api.users.namespace import api as users_api
+from app.api.search.namespace import api as search_api
+from app.api.reporting.namespace import api as reporting_api
+from app.api.variances.namespace import api as variances_api
+from app.api.incidents.namespace import api as incidents_api
+from app.api.now_submissions.namespace import api as now_api
 
 from app.commands import register_commands
 from app.config import Config
@@ -69,6 +70,7 @@ def register_routes(app):
     # Set URL rules for resources
     app.add_url_rule('/', endpoint='index')
 
+    api.add_namespace(compliance_api)
     api.add_namespace(mines_api)
     api.add_namespace(parties_api)
     api.add_namespace(download_token_api)
