+++ python-backend/app/api/download_token/namespace.py
@@ -1,6 +1,6 @@
 from flask_restplus import Namespace
 
-from app.api.download_token.resources.download_token import DownloadTokenResource
+from .resources.download_token import DownloadTokenResource
 
 api = Namespace('download-token', description='A for downloading files from the Document Manager.')
