+++ python-backend/app/api/mines/compliance/resources/compliance.py
@@ -12,17 +12,7 @@ from app.api.services import NRIS_API_service
 from app.extensions import cache
 from ..response_models import MINE_COMPLIANCE_RESPONSE_MODEL
 
-from app.api.mines.mine.models.mine import Mine
-
-
-class DateTime(fields.Raw):
-    def format(self, value):
-        return value.strftime("%Y-%m-%d %H:%M") if value else None
-
-
-class Date(fields.Raw):
-    def format(self, value):
-        return value.strftime("%Y-%m-%d") if value else None
+from ...mine.models.mine import Mine
 
 
 class MineComplianceSummaryResource(Resource, UserMixin):
