+++ python-backend/app/api/mines/namespace.py
@@ -1,37 +1,36 @@
 from flask_restplus import Namespace
 
-from ..mine.resources.mine_map import MineMapResource
-from ..mine.resources.mine import MineResource, MineListSearch, MineListResource
-from ..mine.resources.mine_type import MineTypeResource, MineTypeListResource
-from ..mine.resources.mine_tenure_type_code import MineTenureTypeCodeResource
-from ..mine.resources.mine_disturbance_code import MineDisturbanceCodeResource
-from ..mine.resources.mine_commodity_code import MineCommodityCodeResource
-from ..status.resources.status import MineStatusXrefListResource
-from ..region.resources.region import MineRegionResource
-from ..tailings.resources.tailings import MineTailingsStorageFacilityListResource
-from ..documents.mines.resources.mine_document_resource import MineDocumentListResource
-from ..compliance.resources.compliance import MineComplianceSummaryResource
-from ..compliance.resources.compliance_article import ComplianceArticleResource
-from ..mine.resources.mine_basicinfo import MineBasicInfoResource
-from ..mine.resources.mine_verified_status import MineVerifiedStatusResource, MineVerifiedStatusListResource
-from ..subscription.resources.subscription import MineSubscriptionResource, MineSubscriptionListResource
-from ..applications.resources.application import ApplicationResource, ApplicationListResource
-from ..applications.resources.application_status_code import ApplicationStatusCodeResource
-from ..variances.resources.variance import MineVarianceResource
-from ..variances.resources.variance_list import MineVarianceListResource
-from ..variances.resources.variance_document_upload import MineVarianceDocumentUploadResource
-from ..variances.resources.variance_uploaded_documents import MineVarianceUploadedDocumentsResource
-from ..incidents.resources.mine_incidents import MineIncidentListResource, MineIncidentResource
-from ..incidents.resources.mine_incident_document import MineIncidentDocumentListResource, MineIncidentDocumentResource
-from ..reports.resources.mine_report_document import MineReportDocumentListResource
-from ..reports.resources.mine_reports import MineReportListResource, MineReportResource
-from ..reports.resources.mine_report_definition import MineReportDefinitionListResource
-from ..reports.resources.mine_report_comment import MineReportCommentResource, MineReportCommentListResource
-from ..permits.permit.resources.permit import PermitResource, PermitListResource
-from ..permits.permit.resources.permit_status_code import PermitStatusCodeResource
-from ..permits.permit_amendment.resources.permit_amendment import PermitAmendmentResource, PermitAmendmentListResource
-from ..permits.permit_amendment.resources.permit_amendment_document import PermitAmendmentDocumentListResource, PermitAmendmentDocumentResource
-from ..now_submissions.resources.application_resource import MineApplicationResource
+from .mine.resources.mine_map import MineMapResource
+from .mine.resources.mine import MineResource, MineListSearch, MineListResource
+from .mine.resources.mine_type import MineTypeResource, MineTypeListResource
+from .mine.resources.mine_tenure_type_code import MineTenureTypeCodeResource
+from .mine.resources.mine_disturbance_code import MineDisturbanceCodeResource
+from .mine.resources.mine_commodity_code import MineCommodityCodeResource
+from .status.resources.status import MineStatusXrefListResource
+from .region.resources.region import MineRegionResource
+from .tailings.resources.tailings import MineTailingsStorageFacilityListResource
+from .documents.mines.resources.mine_document_resource import MineDocumentListResource
+from .compliance.resources.compliance import MineComplianceSummaryResource
+from .mine.resources.mine_basicinfo import MineBasicInfoResource
+from .mine.resources.mine_verified_status import MineVerifiedStatusResource, MineVerifiedStatusListResource
+from .subscription.resources.subscription import MineSubscriptionResource, MineSubscriptionListResource
+from .applications.resources.application import ApplicationResource, ApplicationListResource
+from .applications.resources.application_status_code import ApplicationStatusCodeResource
+from .variances.resources.variance import MineVarianceResource
+from .variances.resources.variance_list import MineVarianceListResource
+from .variances.resources.variance_document_upload import MineVarianceDocumentUploadResource
+from .variances.resources.variance_uploaded_documents import MineVarianceUploadedDocumentsResource
+from .incidents.resources.mine_incidents import MineIncidentListResource, MineIncidentResource
+from .incidents.resources.mine_incident_document import MineIncidentDocumentListResource, MineIncidentDocumentResource
+from .reports.resources.mine_report_document import MineReportDocumentListResource
+from .reports.resources.mine_reports import MineReportListResource, MineReportResource
+from .reports.resources.mine_report_definition import MineReportDefinitionListResource
+from .reports.resources.mine_report_comment import MineReportCommentResource, MineReportCommentListResource
+from .permits.permit.resources.permit import PermitResource, PermitListResource
+from .permits.permit.resources.permit_status_code import PermitStatusCodeResource
+from .permits.permit_amendment.resources.permit_amendment import PermitAmendmentResource, PermitAmendmentListResource
+from .permits.permit_amendment.resources.permit_amendment_document import PermitAmendmentDocumentListResource, PermitAmendmentDocumentResource
+from .now_submissions.resources.application_resource import MineApplicationResource
 from app.api.parties.party_appt.resources.mine_party_appt_document_upload_resource import MinePartyApptDocumentUploadResource
 
 api = Namespace('mines', description='Mine related operations')
@@ -51,7 +50,6 @@ api.add_resource(MineTailingsStorageFacilityListResource, '/<string:mine_guid>/t
 api.add_resource(MineDocumentListResource, '/<string:mine_guid>/documents')
 
 api.add_resource(MineComplianceSummaryResource, '/<string:mine_no>/compliance/summary')
-api.add_resource(ComplianceArticleResource, '/compliance/codes')
 
 api.add_resource(MineTypeResource, '/<string:mine_guid>/mine-types/<string:mine_type_guid>')
 api.add_resource(MineTypeListResource, '/<string:mine_guid>/mine-types')
