+++ python-backend/app/api/parties/namespace.py
@@ -1,10 +1,10 @@
 from flask_restplus import Namespace
 
-from ..party.resources.party_resource import PartyResource
-from ..party.resources.sub_division_code_resource import SubDivisionCodeResource
-from ..party_appt.resources.mine_party_appt_resource import MinePartyApptResource
-from ..party_appt.resources.mine_party_appt_type_resource import MinePartyApptTypeResource
-from ..party.resources.party_list_resource import PartyListResource
+from .party.resources.party_resource import PartyResource
+from .party.resources.sub_division_code_resource import SubDivisionCodeResource
+from .party_appt.resources.mine_party_appt_resource import MinePartyApptResource
+from .party_appt.resources.mine_party_appt_type_resource import MinePartyApptTypeResource
+from .party.resources.party_list_resource import PartyListResource
 
 api = Namespace('parties', description='Party related operations')
