+++ python-backend/app/api/reporting/namespace.py
@@ -1,6 +1,6 @@
 from flask_restplus import Namespace
 
-from ..resources.metabase import MetabaseDashboardResource
+from .resources.metabase import MetabaseDashboardResource
 
 api = Namespace('reporting', description='Authenticated reports')
