+++ python-backend/app/api/variances/namespace.py
@@ -1,8 +1,8 @@
 from flask_restplus import Namespace
 
-from ..resources.variance_resource import VarianceResource
-from ..resources.variance_application_status_code import VarianceApplicationStatusCodeResource
-from ..resources.variance_document_category_code import VarianceDocumentCategoryCodeResource
+from .resources.variance_resource import VarianceResource
+from .resources.variance_application_status_code import VarianceApplicationStatusCodeResource
+from .resources.variance_document_category_code import VarianceDocumentCategoryCodeResource
 
 api = Namespace('variances', description='Variances actions/options')
