+++ python-backend/tests/auth/test_expected_auth.py
@@ -4,7 +4,7 @@ from app.api.utils.access_decorators import VIEW_ALL, MINE_EDIT, MINE_ADMIN, MIN
 from app.api.download_token.resources.download_token import DownloadTokenResource
 from app.api.mines.documents.mines.resources.mine_document_resource import MineDocumentListResource
 from app.api.mines.compliance.resources.compliance import MineComplianceSummaryResource
-from app.api.mines.compliance.resources.compliance_article import ComplianceArticleResource
+from app.api.compliance.resources.compliance_article import ComplianceArticleResource
 from app.api.mines.mine.resources.mine_commodity_code import MineCommodityCodeResource
 from app.api.mines.mine.resources.mine_disturbance_code import MineDisturbanceCodeResource
 from app.api.mines.mine.resources.mine_tenure_type_code import MineTenureTypeCodeResource
