+++ python-backend/tests/mines/mine/models/test_mine_region_model.py
