+++ python-backend/tests/mines/mine/resources/test_mine_basicinfo_resource.py
