+++ python-backend/tests/mines/mine/resources/test_mine_commodity_code.py
