+++ python-backend/tests/mines/mine/resources/test_mine_list_by_name_resource.py
