+++ python-backend/tests/mines/mine/resources/test_mine_resource.py
