+++ python-backend/tests/mines/mine/resources/test_mine_type_resource.py
