+++ python-backend/tests/mines/mine/resources/test_mine_verified_status_resource.py
