+++ python-backend/tests/mines/permit/models/test_permit_model.py
