+++ python-backend/tests/mines/permit/resources/test_permit_resource.py
